requirements.txt
django==2.2.9
django-export-action==0.1.1
pillow
djangocms_admin_style==3.3.1
mysqlclient==1.1.0
django-crispy-forms==1.8.0
xlwt==1.3.0
requests==2.31.0
